requirements.txt
streamlit==1.46.0
boto3>=1.34.0
sagemaker==2.190.0
pillow==10.0.0
pandas==2.0.0
numpy==1.24.0
python-dotenv==1.1.1